fix: join test classes
